+++ tests/test_matches.py
@@ -1,7 +1,7 @@
 import pytest
+from async_asgi_testclient import TestClient
 from dirty_equals import IsTrueLike, IsUUID
 from fastapi import status
-from httpx import AsyncClient
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.auth.models import AuthUser
@@ -21,7 +21,7 @@ class TestMatch:
     async def test_access_not_authenticated_matches_list(
         self,
         user2: AuthUser,
-        async_client: AsyncClient,
+        async_client: TestClient,
     ):
         """Проверка существования эндпоинта matches/ и наличия
         доступа к нему неавторизованного пользователя
@@ -40,7 +40,7 @@ class TestMatch:
     async def test_matches_list(
         self,
         authorised_cookie: dict,
-        async_client: AsyncClient,
+        async_client: TestClient,
         user: AuthUser,
         user2: AuthUser,
         match: Match,
@@ -102,7 +102,7 @@ class TestMatch:
 
     async def test_access_not_authenticated_match_delete(
         self,
-        async_client: AsyncClient,
+        async_client: TestClient,
         user2: AuthUser,
         match: Match,
     ):
@@ -122,7 +122,7 @@ class TestMatch:
 
     async def test_valid_match_delete(
         self,
-        async_client: AsyncClient,
+        async_client: TestClient,
         user: AuthUser,
         match: Match,
         authorised_cookie: dict,
@@ -164,7 +164,7 @@ class TestMatch:
     @pytest.mark.skip()
     async def test_invalid_match_delete(
         self,
-        async_client: AsyncClient,
+        async_client: TestClient,
         user2: AuthUser,
         match1: Match,
         authorised_cookie: dict,
